+++ Changelog.md
@@ -3,6 +3,7 @@
 Breaking Changes:
 
 * Ruby < 2.3 is no longer supported. (Phil Pirozhkov, #1349)
+* Extract monkey-patching `should_receive`/`stub` syntax. (Phil Pirozhkov, #1365)
 
 Bug Fixes:
