+++ features/.nav
@@ -32,12 +32,6 @@
 - working_with_legacy_code:
   - any_instance.feature
   - message_chains.feature
-- old_syntax:
-  - stub.feature
-  - should_receive.feature
-  - any_instance.feature
-  - stub_chain.feature
-  - unstub.feature
 - outside_rspec:
   - minitest.feature
   - any_test_framework.feature
